[#PE-260][#PE-261] Changed categories and provided validation to discount publishing

+++ src/main/java/it/gov/pagopa/cgn/portal/service/BucketService.java
@@ -168,7 +168,8 @@ public class BucketService {
 
     @Transactional(Transactional.TxType.REQUIRES_NEW)
     public void deleteBucketCodes(Long discountId) {
-        discountBucketCodeRepository.deleteByDiscountId(discountId);
+        if (discountId != null)
+            discountBucketCodeRepository.deleteByDiscountId(discountId);
     }
 
     public Long countLoadedCodes(DiscountEntity discountEntity) {
